backend/requirements.txt
anthropic>=0.40.0
flask>=3.0.0
flask-socketio>=5.3.6
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
eventlet>=0.35.0
